Refactor [Golang] File System
- [+] refactor(file_system.go): rename package from "main" to "filesystem" and move file to "filesystem" directory
- [+] refactor(file_system_mock.go): rename package from "main" to "filesystem" and move file to "filesystem" directory

Note: This is a better approach because not everything is contained within the "main" package.
